Rename setup package to manual
