+++ WordPress/src/main/java/org/wordpress/android/modules/ApplicationModule.java
@@ -29,7 +29,7 @@ import org.wordpress.android.ui.stats.refresh.lists.widget.configuration.StatsWi
 import org.wordpress.android.ui.stats.refresh.lists.widget.configuration.StatsWidgetDataTypeSelectionDialogFragment;
 import org.wordpress.android.ui.stats.refresh.lists.widget.configuration.StatsWidgetSiteSelectionDialogFragment;
 import org.wordpress.android.ui.stats.refresh.lists.widget.minified.StatsMinifiedWidgetConfigureFragment;
-import org.wordpress.android.util.config.setup.ManualFeatureConfigFragment;
+import org.wordpress.android.util.config.manual.ManualFeatureConfigFragment;
 import org.wordpress.android.util.wizard.WizardManager;
 import org.wordpress.android.viewmodel.gif.provider.GifProvider;
 import org.wordpress.android.viewmodel.gif.provider.TenorProvider;
