+++ WordPress/src/main/java/org/wordpress/android/util/config/manual/ManualFeatureConfigFragment.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.util.config.setup
+package org.wordpress.android.util.config.manual
 
 import android.os.Bundle
 import android.view.LayoutInflater
